refactor(log-viewer): extract RewrittenData into testable module
Move the RewrittenData class out of scripts.js (which touches the DOM at
load and can't be imported in tests) into rewrite.js, and add unit tests
covering the indexer non-array changes regression and array/absent cases.
Behavior is unchanged.
